implementing custom exceptions

+++ src/main/java/br/com/fiap/foodlivery/foodlivery/services/UsuarioService.java
@@ -3,9 +3,12 @@ package br.com.fiap.foodlivery.foodlivery.services;
 import br.com.fiap.foodlivery.foodlivery.dtos.UsuarioRequestDTO;
 import br.com.fiap.foodlivery.foodlivery.dtos.UsuarioResponseDTO;
 import br.com.fiap.foodlivery.foodlivery.entities.Usuario;
+import br.com.fiap.foodlivery.foodlivery.exceptions.custom.DatabaseException;
+import br.com.fiap.foodlivery.foodlivery.exceptions.custom.ResourceNotFoundException;
 import br.com.fiap.foodlivery.foodlivery.mappers.UsuarioMapper;
 import br.com.fiap.foodlivery.foodlivery.repositories.UsuarioRepository;
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -28,7 +31,7 @@ public class UsuarioService {
 
     public UsuarioResponseDTO findById(Long id) {
         Usuario usuario = usuarioRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
 
         return usuarioMapper.toResponseDTO(usuario);
     }
@@ -40,16 +43,29 @@ public class UsuarioService {
     }
 
     public UsuarioResponseDTO update(Long id, UsuarioRequestDTO usuarioRequestDTO) {
-        Usuario existingUsuario = usuarioRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado"));
-        existingUsuario = usuarioMapper.updateDromDTO(usuarioRequestDTO, existingUsuario);
-        existingUsuario = usuarioRepository.save(existingUsuario);
-        UsuarioResponseDTO usuarioResponseDTO = usuarioMapper.toResponseDTO(existingUsuario);
-        return usuarioResponseDTO;
+
+        try {
+            Usuario existingUsuario = usuarioRepository.getReferenceById(id);
+            existingUsuario = usuarioMapper.updateFromDTO(usuarioRequestDTO, existingUsuario);
+            existingUsuario = usuarioRepository.save(existingUsuario);
+            UsuarioResponseDTO usuarioResponseDTO = usuarioMapper.toResponseDTO(existingUsuario);
+            return usuarioResponseDTO;
+
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     public void delete(Long id) {
-        usuarioRepository.deleteById(id);
+
+        if (!usuarioRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            usuarioRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
 }
